+++ patrol/sentry_api_client/resource.py
@@ -3,6 +3,7 @@ from simple_rest_client.resource import Resource
 
 class ProjectEvents(Resource):
     actions = {
+        'fetch': {'method': 'GET', 'url': 'projects/{}/{}/events/{}'},
         'list': {'method': 'GET', 'url': 'projects/{}/{}/events/'}
     }
 
@@ -10,4 +11,11 @@ class ProjectEvents(Resource):
 class ProjectIssues(Resource):
     actions = {
         'list': {'method': 'GET', 'url': 'projects/{}/{}/issues/'}
+    }
+
+
+class Issues(Resource):
+    actions = {
+        'fetch': {'method': 'GET', 'url': 'issues/{}/'},
+        'update': {'method': 'PUT', 'url': 'issues/{}/'},
     }